Ititial, confused code
Two pieces of stolen example code for their corresponding libraries, are used to communicate with the ADC and loadcell-amplifier development boards respectively.

The ledc functions are used to generate a PWM signal which, hopefully, will instruct the electronic speed controller to start up the motor (at 0.5% duty cycle / 1ms signal width in a PWM frequency of 50Hz / 20ms), and then ramp up its speed to 50% (0.75% duty / 1.5ms) with a gradual, linear increase of 0.5 percentage points per every 1/163 seconds. This should give a ramp up from 0 to 50 % in 10 seconds. Data is read and sent to the serial console every second during the motor ramp up.

 When the ramp up is complete, data continues to be sent every second and motor rpm is held at a constant 50%.

NOTE: There currently is nothing ending the programme, and no emergency stop. That is bad as it is supposed to drive a motor. Therefore the program should not be run with a motor attached, and must be stopped by force.

NOTE: I have not tested the programme, I have not tried to compile it, I have not proofread it, and I have not checked the math.