proper voltage, current and loadcell value readout
the values are corrected by measuring the voltage or current with a multimeter of doubtable quality and noting the value given by the adc and then dividing the wanted value, measured by the multimeter, by the given value. The resulting decimal number is then input into haubergs raction aproximator https://haubergs.com/fractions. (This was the first tool I found.) The resulting fraction is then multiplied with the adc value being corrected.

I also checked that the calibration factor worked at different voltages and currents to make sure that the adc's response is somewhat linear.

This method is obviously not ideal, and pretty stupid, but results in the value only having to be divided once, which makes this horribly frankensteined code look at least a little nicer.