updated value of "Calibration for current measurement updated."
This time the value is taken for actually putting a few (20) values in a table, and getting the value through linear regression.

The previous value was inaccurate as the multimeter used to measure the actual current, updates quite a bit slower than the program on the ESP 32. Therefor when arbitrary value it i easy to choose a multimeter reading that shows the current-value that is no longer accurate.